Refactor: Relocate UI components and update dependencies
Moves `ActorItem`, `CategoryItem`, `HomeCard`, and `ImageView` components from the `designSystem` module to the `presentation` module's `shared` package.

Updates import statements across various screens in the `presentation` module to reflect the new location of these components.

Removes `imageHaramBlur` dependency from `designSystem` and adds it to the `presentation` module.
