+++ presentation/src/main/java/com/london/presentation/feature/continuewatching/ContinueWatchingScreen.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.london.designsystem.R
-import com.london.designsystem.component.HomeCard
+import com.london.presentation.shared.HomeCard
 import com.london.designsystem.component.NovixChip
 import com.london.designsystem.component.TabItem
 import com.london.designsystem.component.TabLayout
