+++ presentation/src/main/java/com/london/presentation/feature/details/actordetails/gallery/ActorGalleryScreen.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.london.designsystem.component.CircularLoading
-import com.london.designsystem.component.ImageView
+import com.london.presentation.shared.ImageView
 import com.london.designsystem.component.TopBar
 import com.london.designsystem.component.button.ErrorImage
 import com.london.designsystem.theme.NovixTheme
