+++ presentation/src/main/java/com/london/presentation/feature/details/tvshow/episodedetails/EpisodeDetailsScreen.kt
@@ -44,7 +44,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.london.designsystem.component.ActorItem
+import com.london.presentation.shared.ActorItem
 import com.london.designsystem.component.Icon
 import com.london.designsystem.component.Text
 import com.london.designsystem.component.TopBar
