+++ presentation/src/main/java/com/london/presentation/feature/details/tvshow/tvshowdetails/TvShowDetailsScreen.kt
@@ -50,10 +50,10 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.london.designsystem.R
-import com.london.designsystem.component.ActorItem
+import com.london.presentation.shared.ActorItem
 import com.london.designsystem.component.CircularLoading
 import com.london.designsystem.component.Icon
-import com.london.designsystem.component.ImageView
+import com.london.presentation.shared.ImageView
 import com.london.designsystem.component.NovixChip
 import com.london.designsystem.component.Text
 import com.london.designsystem.component.TopBar
