+++ presentation/src/main/java/com/london/presentation/feature/reviews/ReviewsScreen.kt
@@ -40,7 +40,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.paging.LoadState
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.london.designsystem.component.CircularLoading
-import com.london.designsystem.component.ImageView
+import com.london.presentation.shared.ImageView
 import com.london.designsystem.component.Text
 import com.london.designsystem.component.TopBar
 import com.london.designsystem.component.button.ErrorImage
