+++ presentation/src/main/java/com/london/presentation/feature/toprated/TopRatedScreen.kt
@@ -31,7 +31,7 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.paging.LoadState
 import androidx.paging.compose.collectAsLazyPagingItems
 import com.london.designsystem.R
-import com.london.designsystem.component.HomeCard
+import com.london.presentation.shared.HomeCard
 import com.london.designsystem.component.NovixChip
 import com.london.designsystem.component.TabItem
 import com.london.designsystem.component.TabLayout
