+++ presentation/src/main/java/com/london/presentation/shared/MoviesLayout.kt
@@ -10,7 +10,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.paging.compose.LazyPagingItems
-import com.london.designsystem.component.HomeCard
 import com.london.domain.entity.Movie
 import com.london.presentation.utils.gridColmuns
